Creating service and custom some exceptions

+++ src/main/java/bj/vargas/vargasfood/infraestructure/repository/KitchenRepositoryImpl.java
@@ -2,6 +2,7 @@ package bj.vargas.vargasfood.infraestructure.repository;
 
 import bj.vargas.vargasfood.domain.model.Kitchen;
 import bj.vargas.vargasfood.domain.repository.KitchenRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 
 import javax.persistence.EntityManager;
@@ -18,24 +19,27 @@ public class KitchenRepositoryImpl implements KitchenRepository {
 
     @Override
     public List<Kitchen> list() {
-        TypedQuery<Kitchen> query = entityManager.createQuery("from Kitchen", Kitchen.class);
+        final TypedQuery<Kitchen> query = entityManager.createQuery("from Kitchen", Kitchen.class);
         return query.getResultList();    }
 
     @Override
-    public Kitchen getKitchen(Long id) {
+    public Kitchen getKitchen(final Long id) {
         return entityManager.find(Kitchen.class, id);
     }
 
     @Override
     @Transactional
-    public Kitchen save(Kitchen kitchen) {
+    public Kitchen save(final Kitchen kitchen) {
         return entityManager.merge(kitchen);
     }
 
     @Override
     @Transactional
-    public void remove(Kitchen kitchen) {
-        kitchen = getKitchen(kitchen.getId());
+    public void remove(final Long id) {
+        final Kitchen kitchen = getKitchen(id);
+        if (kitchen == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(kitchen);
     }
 }
